Change in GetUser Service
 Making GetUser service more simple  and moving "Services" folder to "Model" folder.
