+++ Areas/Admin/Controllers/UsersController.cs
@@ -3,26 +3,25 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using HorizonHotelWebsite.Services.GetUsers;
+
+using HorizonHotelWebsite.Models.Services.GetUser;
 
 namespace HorizonHotelWebsite.Areas.Admin.Controllers
 {
     public class UsersController : Controller
     {
-        private readonly IGetUsersService _getUsersService;
-        public UsersController(IGetUsersService getUsersService)
+        private readonly IGetUserService _getUserService;
+        public UsersController(IGetUserService getUserService)
         {
-            _getUsersService = getUsersService;
+            _getUserService = getUserService;
             
         }
         [Area("Admin")]
-        public IActionResult Index(string searchkey)
+        public IActionResult Index()
         {
-            return View(_getUsersService.Execute(new RequestGetUserDto
-            {
-               
-                SearchKey = searchkey,
-            }));
+            
+            return View(_getUserService.Execute());
         }
+       
     }
 }
