+++ Areas/Admin/Views/Shared/_AdminLayout.cshtml
@@ -82,13 +82,13 @@
                         <p class="d-none">Full Screen</p>
                     </a><a class="mr-2 display-inline-block">
                         <i class="ft-shopping-cart blue-grey darken-4"></i>
-                        @*<p class="d-none">سبد خرید</p>*@
+                        <p class="d-none">سبد خرید</p>
                     </a>
                     <div class="dropdown mr-2 display-inline-block">
-                        <a id="apps" href="#" data-toggle="dropdown" class="nav-link position-relative dropdown-toggle"><i class="ft-edit blue-grey darken-4"></i><span class="mx-1 blue-grey darken-4 text-bold-400">@*برنامه ها*@</span></a>
-                        @*<div class="apps dropdown-menu">
+                        <a id="apps" href="#" data-toggle="dropdown" class="nav-link position-relative dropdown-toggle"><i class="ft-edit blue-grey darken-4"></i><span class="mx-1 blue-grey darken-4 text-bold-400">برنامه ها</span></a>
+                        <div class="apps dropdown-menu">
                             <div class="arrow_box"><a href="chat.html" class="dropdown-item py-1"><span>چت</span></a><a href="taskboard.html" class="dropdown-item py-1"><span>وظیفه</span></a><a href="calendar.html" class="dropdown-item py-1"><span>تقویم</span></a></div>
-                        </div>*@
+                        </div>
                     </div>
                 </div>
                 <div class="navbar-container">
