+++ Areas/Admin/Views/Users/Index.cshtml
@@ -1,5 +1,5 @@
-﻿@using HorizonHotelWebsite.Services.GetUsers;
-@model ReslutGetUserDto;
+﻿@using HorizonHotelWebsite.Models.Services.GetUser;
+@model ResultGetUser
 @{
     ViewData["Title"] = "Index";
     Layout = "~/Areas/Admin/Views/Shared/_AdminLayout.cshtml";
