+++ Models/Services/GetUser/IGetUserService.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace HorizonHotelWebsite.Models.Services.GetUser
+{
+    public interface IGetUserService
+    {
+        ResultGetUser Execute();
+    }
+}
