+++ Models/Services/GetUser/ResultGetUser.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace HorizonHotelWebsite.Models.Services.GetUser
+{
+    public class ResultGetUser
+    {
+        public List<GetUserView> Users { get; set; }
+    }
+}
