--- Services/GetUsers/IGetUsersService.cs
@@ -1,16 +1,0 @@
-﻿
-
-using HorizonHotelWebsite.Services.Interfaces;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace HorizonHotelWebsite.Services.GetUsers
-{
-    public interface IGetUsersService
-    {
-        
-        ReslutGetUserDto Execute(RequestGetUserDto requestGetUserDto);
-    }
-}
