--- Services/GetUsers/ReslutGetUserDto.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace HorizonHotelWebsite.Services.GetUsers
-{
-    public class ReslutGetUserDto
-    {
-        public List<GetUsersDto> Users { get; set; }
-    }
-}
