+++ appsettings.Development.json
@@ -6,8 +6,7 @@
             "Microsoft.Hosting.Lifetime": "Information"
         }
     },
-    //"ConnectionStrings": {
-    //    "DataBaseContext": "Server=(localdb)\\mssqllocaldb;Database=TheLocalDatabase-1;Trusted_Connection=True;MultipleActiveResultSets=true"
+    
 
     "ConnectionStrings": {
         "DataBaseContext": "Server=(localdb)\\mssqllocaldb;Database=HorizonHotelWebsite;Trusted_Connection=True;MultipleActiveResultSets=True"
